requirement.txt
faiss-cpu
fastapi==0.75.0
fonttools==4.25.0
h11==0.13.0
httptools==0.2.0
idna==3.3
inflection==0.5.1
itsdangerous==2.1.1
numpy==1.19.5
opencv-python==4.5.5.64
orjson==3.6.7
paho-mqtt==1.6.1
pandas==1.3.4
parso==0.8.3
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
Pillow==8.3.1
pkginfo==1.7.1
ply==3.11
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==7.0.0
pycosat==0.6.3
pycparser==2.21
pycurl==7.44.1
pydantic==1.9.0
pydeck==0.7.1
pyls-spyder==0.4.0
pyodbc===4.0.0-unsupported
pyparsing==3.0.7
pyserial==3.5
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.18.3
simplegeneric==0.8.1
sip==4.19.13
smmap==5.0.0
Sphinx==4.2.0
starlette==0.17.1
streamlit==0.86.0
TBB==0.2
terminado==0.9.4
timm==0.4.12
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
ujson==4.3.0
unicodecsv==0.14.1
urllib3==1.26.9
uvicorn==0.15.0
wget==3.2
